2025 day 05: part 2 🥵
Trick was that one new fresh range may overlap / extend several existing
ranges. So that's why all my initial solutions did not work, because of
this case. The overlap cases were right, the merge was right, the size
was right... only my input included such a case. That was super tricky
to find out.